+++ src/assets/index.ts
@@ -1,4 +1,5 @@
-//mainexport { default as MainHot } from './main/mainHot.svg';
+// main
+export { default as MainHot } from './main/mainHot.svg';
 export { default as MainBonfire } from './main/mainBonfire.svg';
 export { default as MainLoudspeaker } from './main/mainLoudspeaker.svg';
